pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deskmaid"
version = "0.1.0"
description = "AI-powered desktop file organizer"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "openai>=1.0.0",
    "python-docx>=1.0.0",
    "openpyxl>=3.1.0",
    "python-pptx>=0.6.21",
    "pdfplumber>=0.9.0",
]

[project.scripts]
deskmaid = "deskmaid.cli:app"

[tool.setuptools.packages.find]
include = ["deskmaid*"]
